feat: validation 로직 도입, 코드 정리
service 계층에서 값 검증을 시도함, 전체적인 코드를 정리하였음

+++ src/main/java/org/programmers/springbootbasic/voucher/FixedDiscountVoucher.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 public class FixedDiscountVoucher implements Voucher {
 
     private final UUID id;
-    private final long amount;
+    private final int amount;
 
     @Override
     public long discount(long beforeDiscount) {
@@ -21,7 +21,7 @@ public class FixedDiscountVoucher implements Voucher {
     }
 
     @Override
-    public long getAmount() {
+    public int getAmount() {
         return amount;
     }
