+++ src/main/java/org/programmers/springbootbasic/voucher/Voucher.java
@@ -5,6 +5,8 @@ import java.util.UUID;
 
 public interface Voucher {
     long discount(long beforeDiscount);
+
     UUID getId();
-    long getAmount();
+
+    int getAmount();
 }
